Modified descriptions and added list/search functionality
'list' with no parameters lists all items in the collection. Given
parameters, it lists all those with the keyword and prints them to the
terminal. 'list' prints to the screen the filenames with metadata which
match the search term, without the filename extensions.

Descriptions of MM_File were changed to match assignment specfications.

Still working on indexing listed results to be used with 'add', 'set',
and 'del' functions.