Added file export functionality. All formats of 'filename' now accepted
File export encodes the MMFile as a JSON and writes out as expected. If
only the filename is given (without full path) then the file is written
to the root directory of the project.

Fixed issue where filename was statically coded to a certain destination
but now it's dynamic. '~' expands to the users home directory, '/' marks
it to be the full path, otherwise it is assumed that the user wants to
read/write to/from the root directory of the project.

Also removed '\' from being encoded into the exported JSON file under
the attribute 'fullpath'. It was previously putting backslashes to
escape the following forward slashes.